refactor: progress bar 를 사용하는 glide 리스너를 하나의 파일로 분리한다
