+++ android/app/src/main/java/poke/rogue/helper/presentation/util/view/CommonBindingAdapter.kt
@@ -1,15 +1,11 @@
 package poke.rogue.helper.presentation.util.view
 
-import android.graphics.drawable.Drawable
 import android.view.View
 import android.widget.ImageView
 import androidx.annotation.ColorRes
 import androidx.core.content.ContextCompat
 import androidx.databinding.BindingAdapter
 import com.bumptech.glide.Glide
-import com.bumptech.glide.load.DataSource
-import com.bumptech.glide.load.engine.GlideException
-import com.bumptech.glide.request.RequestListener
 import com.google.android.material.progressindicator.CircularProgressIndicator
 import poke.rogue.helper.R
 import poke.rogue.helper.presentation.util.context.colorOf
@@ -35,37 +31,14 @@ fun ImageView.setCroppedImage(imageUrl: String?) {
 
 @BindingAdapter("imageUrlWithProgress", "progressIndicator")
 fun ImageView.loadImageWithProgress(
-    url: String?,
+    imageUrl: String?,
     progressIndicator: CircularProgressIndicator,
 ) {
     progressIndicator.visibility = View.VISIBLE
 
     Glide.with(context)
-        .load(url)
-        .listener(
-            object : RequestListener<Drawable> {
-                override fun onLoadFailed(
-                    e: GlideException?,
-                    model: Any?,
-                    target: com.bumptech.glide.request.target.Target<Drawable>?,
-                    isFirstResource: Boolean,
-                ): Boolean {
-                    progressIndicator.visibility = View.GONE
-                    return false
-                }
-
-                override fun onResourceReady(
-                    resource: Drawable?,
-                    model: Any?,
-                    target: com.bumptech.glide.request.target.Target<Drawable>?,
-                    dataSource: DataSource?,
-                    isFirstResource: Boolean,
-                ): Boolean {
-                    progressIndicator.visibility = View.GONE
-                    return false
-                }
-            },
-        )
+        .load(imageUrl)
+        .listener(createProgressListener(progressIndicator))
         .error(R.drawable.icon_error)
         .into(this)
 }
@@ -79,30 +52,7 @@ fun ImageView.setCroppedImageWithProgressIndicator(
 
     Glide.with(context)
         .load(imageUrl)
-        .listener(
-            object : RequestListener<Drawable> {
-                override fun onLoadFailed(
-                    e: GlideException?,
-                    model: Any?,
-                    target: com.bumptech.glide.request.target.Target<Drawable>?,
-                    isFirstResource: Boolean,
-                ): Boolean {
-                    progressIndicator.visibility = View.GONE
-                    return false
-                }
-
-                override fun onResourceReady(
-                    resource: Drawable?,
-                    model: Any?,
-                    target: com.bumptech.glide.request.target.Target<Drawable>?,
-                    dataSource: DataSource?,
-                    isFirstResource: Boolean,
-                ): Boolean {
-                    progressIndicator.visibility = View.GONE
-                    return false
-                }
-            },
-        )
+        .listener(createProgressListener(progressIndicator))
         .error(R.drawable.icon_error)
         .transform(CropMarginTransformation())
         .into(this)
