+++ android/app/src/main/java/poke/rogue/helper/presentation/util/view/ProgressListener.kt
@@ -1,0 +1,33 @@
+package poke.rogue.helper.presentation.util.view
+
+import android.graphics.drawable.Drawable
+import android.view.View
+import com.bumptech.glide.load.DataSource
+import com.bumptech.glide.load.engine.GlideException
+import com.bumptech.glide.request.RequestListener
+import com.bumptech.glide.request.target.Target
+import com.google.android.material.progressindicator.CircularProgressIndicator
+
+fun createProgressListener(progressIndicator: CircularProgressIndicator): RequestListener<Drawable> =
+    object : RequestListener<Drawable> {
+        override fun onLoadFailed(
+            e: GlideException?,
+            model: Any?,
+            target: Target<Drawable>?,
+            isFirstResource: Boolean,
+        ): Boolean {
+            progressIndicator.visibility = View.GONE
+            return false
+        }
+
+        override fun onResourceReady(
+            resource: Drawable?,
+            model: Any?,
+            target: Target<Drawable>?,
+            dataSource: DataSource?,
+            isFirstResource: Boolean,
+        ): Boolean {
+            progressIndicator.visibility = View.GONE
+            return false
+        }
+    }
